Refactored to move the ExecutionGraph to its own module.
